+++ firebase-sessions/src/main/kotlin/com/google/firebase/sessions/SessionFirelogPublisher.kt
@@ -29,14 +29,14 @@ import kotlinx.coroutines.launch
 import kotlinx.coroutines.tasks.await
 
 /** Responsible for uploading session events to Firelog. */
-internal interface SessionFirelogPublisher {
+internal fun interface SessionFirelogPublisher {
 
   /** Asynchronously logs the session represented by the given [SessionDetails] to Firelog. */
-  fun logSession(sessionDetails: SessionDetails): Unit
+  fun logSession(sessionDetails: SessionDetails)
 
   companion object {
     val instance: SessionFirelogPublisher
-      get() = Firebase.app.get(SessionFirelogPublisher::class.java)
+      get() = Firebase.app[SessionFirelogPublisher::class.java]
   }
 }