+++ firebase-sessions/src/main/kotlin/com/google/firebase/sessions/SessionLifecycleServiceBinder.kt
@@ -26,16 +26,16 @@ import com.google.firebase.FirebaseApp
 import com.google.firebase.app
 
 /** Interface for binding with the [SessionLifecycleService]. */
-internal interface SessionLifecycleServiceBinder {
+internal fun interface SessionLifecycleServiceBinder {
   /**
    * Binds the given client callback [Messenger] to the [SessionLifecycleService]. The given
    * callback will be used to relay session updates to this client.
    */
-  fun bindToService(callback: Messenger, serviceConnection: ServiceConnection): Unit
+  fun bindToService(callback: Messenger, serviceConnection: ServiceConnection)
 
   companion object {
     val instance: SessionLifecycleServiceBinder
-      get() = Firebase.app.get(SessionLifecycleServiceBinder::class.java)
+      get() = Firebase.app[SessionLifecycleServiceBinder::class.java]
   }
 }